requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
fastapi
uvicorn
langgraph
langfuse
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-anthropic
langchain-huggingface
langchain-qdrant
langchain-tavily
langchain-text-splitters
qdrant-client
sentence-transformers
rank-bm25
pdfplumber
peft
pydantic
pydantic-settings
python-dotenv
tqdm
beautifulsoup4
ragas
FlagEmbedding==1.2.11
transformers==4.44.2
